Remove redundant duration capability.
This was intended along with start_offset to allow users to set a subset of an animation to be used by a player within a set of animations being playerd. To reduce confusion duration and offsetting is left to the player and animations instances can no longer be started from point other then their 0 or finish before their natural duration.